refactor: common 파일 구조 변경 #25
