+++ src/common/apis/env.d.ts
