+++ src/domain/rooms/hooks/mutations/roomMutations.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { roomService } from "../../services/roomService";
-import QUERY_KEYS from "../../../../common/hooks/queries/queryKeys";
+import QUERY_KEYS from "../../../../common/constants/queryKeys";
 import { Room } from "../../types/room";
 
 // 방 알림 설정 토글 mutation
